xyz/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.0
google-cloud-aiplatform[all]>=1.60.0
google-cloud-bigquery>=3.25.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.18.0
anthropic>=0.34.0
google-adk>=0.4.0
google-auth>=2.40.3
google-cloud-logging>=3.12.1
opentelemetry-exporter-gcp-logging>=1.9.0a0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation-google-genai>=0.4b0
opentelemetry-instrumentation-logging>=0.58b0
